pyproject.toml
[tool.poetry]
name = "djangobigprojecttemplate"
version = "0.1.0"
description = ""
authors = ["BaseKodu <phenyoitumeleng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-split-settings = "^1.3.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
